moved cWindow stuff calculating prev & next links to PlaylistHeader where they're actually displayed
